+++ crates/api-types/src/relayer.rs
@@ -13,7 +13,7 @@ pub struct PollResult {
     pub max_block_number: u64,
     /// Nonce from the observed source (used for version comparison in JWKManager)
     /// For blockchain events: GravityPortal.MessageSent nonce
-    pub nonce: Option<u64>,
+    pub nonce: Option<u128>,
     /// Whether the state was updated in this poll
     pub updated: bool,
 }
